move MapContext to its own file
